errcode: fix 32-bit loff_t truncation crash in bch2_err_class(), fix memory leak in percpu.c
On 32-bit architectures, `long` is 32-bit while `loff_t` is 64-bit. When
seeking past 2GB (`0x80000000`), passing a 64-bit `loff_t` offset into
`bch2_err_class(long err)` truncates it to `-2147483648` (`INT_MIN`).

`bch2_err_class()` misidentifies this positive offset as a negative error
code (`err < 0`), passing `INT_MIN` to `__bch2_err_class()` and triggering
a kernel panic (`BUG_ON` at `errcode.c:58`).

Fix this by changing `bch2_err_class()` to accept and return `loff_t`,
preserving 64-bit file offsets on 32-bit systems.

Upgrade size_at_grain[] element type from u16 to uint32_t. This fixes a
latent bug where per-CPU allocations > 512KB (65535 * 8 bytes) overflowed
the 16-bit tracker, causing silently leaked upon free_percpu().

+++ linux/percpu.c
@@ -96,7 +96,7 @@ static size_t		dynamic_used;
  * Map from grain index to allocation size in grains, so free_percpu()
  * doesn't need a size argument.
  */
-static u16		size_at_grain[BCH_PERCPU_DYNAMIC_SIZE / BCH_PERCPU_GRAIN];
+static uint32_t		size_at_grain[BCH_PERCPU_DYNAMIC_SIZE / BCH_PERCPU_GRAIN];
 
 static pthread_mutex_t	bch_percpu_lock = PTHREAD_MUTEX_INITIALIZER;
